Rename TextUtils to StringUtils
