+++ collect_app/src/main/java/org/odk/collect/android/widgets/SpinnerMultiWidget.java
@@ -28,6 +28,7 @@ import org.javarosa.core.model.data.SelectMultiData;
 import org.javarosa.core.model.data.helper.Selection;
 import org.odk.collect.android.R;
 import org.odk.collect.android.formentry.questions.QuestionDetails;
+import org.odk.collect.android.utilities.StringUtils;
 import org.odk.collect.android.widgets.interfaces.ButtonWidget;
 import org.odk.collect.android.widgets.interfaces.MultiChoiceWidget;
 import org.odk.collect.android.widgets.warnings.SpacesInUnderlyingValuesWarning;
@@ -78,7 +79,7 @@ public class SpinnerMultiWidget extends ItemsWidget implements ButtonWidget, Mul
         // Build View
         for (int i = 0; i < items.size(); i++) {
             answerItems[i] = questionDetails.getPrompt().getSelectChoiceText(items.get(i));
-            styledAnswerItems[i] = org.odk.collect.android.utilities.TextUtils.textToHtml(answerItems[i]);
+            styledAnswerItems[i] = StringUtils.textToHtml(answerItems[i]);
         }
 
         selectionText = getAnswerTextView();
@@ -183,7 +184,7 @@ public class SpinnerMultiWidget extends ItemsWidget implements ButtonWidget, Mul
         if (selectedValues.isEmpty()) {
             clearAnswer();
         } else {
-            CharSequence answerText = org.odk.collect.android.utilities.TextUtils.textToHtml(String.format(getContext().getString(R.string.selected_answer),
+            CharSequence answerText = StringUtils.textToHtml(String.format(getContext().getString(R.string.selected_answer),
                     TextUtils.join(", ", selectedValues)));
             selectionText.setText(answerText);
             selectionText.setVisibility(View.VISIBLE);
